\name{reportAnnotation}
\alias{reportAnnotation}
\alias{reportStatistics}
\alias{reportGeneLists}
\alias{reportNetworks}
\alias{reportScreenData}
\alias{reportDoublePerturbation}
\alias{reportMainEffects}
\alias{reportHeatmap}

\title{Specialized report functions}

\description{
Functions that provide a html report of genetic interactions screens for specific topics.
}
\usage{
reportAnnotation        (sgi, verbose = 0, path = ".", dir = "annotation",
                         prefix = "annotation", report = NULL)
reportStatistics        (sgi, verbose = 0, path = ".", dir = "stats",
			 prefix = "stat", report = NULL)
reportGeneLists         (sgi, verbose = 0, path = ".", dir = "hitlist", 
			 prefix = "hitlist", report = NULL)
reportNetworks          (sgi, verbose = 0, path = ".", dir = "networks",
			 prefix = "networks", Networks, qv = 0.05,
			 withoutgroups = c("pos", "neg"), report = NULL)
reportScreenData        (sgi, type = "data", design = "template",
			 do.trafo = TRUE, do.inv.trafo = FALSE, verbose = 0,
 			 path = ".", dir = "screenplots", prefix = "screenplot",
			 png.args = list(width = 960, height = 960),
			 pdf.args = list(width = 7, height = 7),
			 plotScreen.args = list(ncol = 6L, do.legend = TRUE,
			                        fill = c("red", "white", "blue")),
			 png.scatter.args = list(width = 400, height = 400),
			 pdf.scatter.args = list(width = 7, height = 7), report = NULL)
reportDoublePerturbation(sgi, verbose = 0, path = ".", dir = "doublePerturbations",
			 prefix = "doublePerturbationPlots", report = NULL,
                         withoutgroups = c("neg", "pos"),
                         png.args = list(width = 500, height = 500),
			 pdf.args = list(width = 7, height = 7), ...)
reportMainEffects       (sgi, verbose = 0, path = ".", dir = "maineffects",
			 prefix = "maineffects",
			 png.args = list(width = 500, height = 500),	
			 pdf.args = list(width = 7, height = 7),
			 plot.args = list(), report = NULL)
reportHeatmap           (sgi, verbose = 0, path = ".", dir = "heatmap",
			 prefix = "heatmap",
			 png.args = list(width = 1000, height = 1000),
			 pdf.args = list(width = 15, height = 15),
			 report = NULL, withoutgroups = c("neg", "pos"))
}
\arguments{
  \item{sgi}{An object of class \code{\link{RNAinteract}}.}
  \item{verbose}{Either 0 (default, no output), 1 (minimum output), or 2 (outout)}
  \item{path}{The main path to the HTML report directory.}
  \item{dir}{A subdirectory where the report is written to.}
  \item{prefix}{A prefix for each file written in the subdirectory. Using different prefixes, one can write multiple reports in the same directory.}
  \item{report}{A report object as generated by \code{\link{startReport}}.}
  \item{Networks}{A boolean array with edges from interaction graphs.}
  \item{qv}{A cut-off value for the q-values.}
  \item{withoutgroups}{Genes annotated with these groups are not plotted in this report.}
  \item{type}{Any "type" that can be passed to \code{\link{getData}}.}
  \item{design}{Either "template" (default) or "query"}
  \item{do.trafo,
do.inv.trafo}{Apply the (inverse) transformation before plotting.}
  \item{png.args}{A list with entries width and height specifying the width and height of the generated png images.}
  \item{pdf.args}{A list with entries width and height specifying the width and height of the generated of files.}
  \item{plotScreen.args}{Arguments for the screen plots}
  \item{png.scatter.args,
pdf.scatter.args,
plot.args}{Arguments for the scatter plots}
  \item{...}{Parameters passed to the plotting functions.}
}
\details{
  Each of these function generates a HTML report. It is added to a HTML frame. 

The report object has to be created with startReport beforehand. Multiple report functions can be called afterwards. When all reports are written, the report is finalized and closed by \code{\link{endReport}}.
}
\value{
All functions return a report object as returned by \code{\link{startReport}}. 
}

\author{
Bernd Fischer
}
\seealso{
\code{\link{RNAinteract-package}},
\code{\link{startReport}}, \code{\link{endReport}}
}
\examples{
data("sgi")
report = startReport("report")
reportAnnotation(sgi, report = report)
endReport(report)
browseURL(file.path("report","index.html"))
}
\keyword{ print }

