\name{sgi}
\alias{sgi}
\docType{data}
\title{
Dataset of class 'RNAinteract'
}
\description{
Sample object of class \code{\link{RNAinteract}}. The data are real but anonymized. The object contains two replicate screens with three readout channels.
}
\usage{data(sgi)}
\format{
	The data contains two screens with three readout channels. The screen is performed on multiwell plates with 8 x 12 wells.
}
\examples{
data(sgi)
sgi
}
\keyword{datasets}
