\name{swaptree}
\alias{swaptree}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Swaps a branch of a hclust object.
}
\description{
Swaps the left and right branch at a specified level of a dendrogram.
}
\usage{
swaptree(hc, level)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{hc}{An hierarchical clustering object as produced by hclust.}
  \item{level}{The level to be swapped.}
}
\value{
Returns an hclust object.
}
\author{
Bernd Fischer
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
